Moved footer outside of the layout component as disclaimers are now on a per page/post basis.
