+++ src/components/Footer/index.tsx
@@ -1,4 +1,4 @@
-import React from "react"
+import React, { FC, ReactNode } from "react"
 import { Link } from "gatsby"
 
 // Styles
@@ -20,10 +20,10 @@ import FooterMenu from "./FooterMenu"
 import PlansByStateMenu from "./PlansByStateMenu";
 
 // Queries
-import { useFooterProductMenuQuery } from "../../hooks/useFooterProductMenuQuery"
-import { useFooterConnectMenuQuery } from "../../hooks/useFooterConnectMenuQuery"
-import { useFooterGetCounselMenuQuery } from "../../hooks/useFooterGetCounselMenuQuery"
-import { useFooterGetToKnowUsMenuQuery } from "../../hooks/useFooterGetToKnowUsMenuQuery"
+import { useFooterProductMenuQuery } from "../../hooks/footer/useFooterProductMenuQuery"
+import { useFooterConnectMenuQuery } from "../../hooks/footer/useFooterConnectMenuQuery"
+import { useFooterGetCounselMenuQuery } from "../../hooks/footer/useFooterGetCounselMenuQuery"
+import { useFooterGetToKnowUsMenuQuery } from "../../hooks/footer/useFooterGetToKnowUsMenuQuery"
 
 // Images
 import Logo from "../../images/HMIA_logo2.png"
@@ -32,7 +32,11 @@ import Facebook from "../../images/facebook.webp"
 import YouTube from "../../images/youtube.webp"
 import BBB from "../../images/bbb-seal.png"
 
-const Footer = () => {
+interface Props {
+  children: ReactNode
+}
+
+const Footer: FC<Props> = ({children}) => {
   const { productMenu } = useFooterProductMenuQuery()
   const { connectMenu } = useFooterConnectMenuQuery()
   const { counselMenu } = useFooterGetCounselMenuQuery()
@@ -69,23 +73,13 @@ const Footer = () => {
           </FooterMenuWrapper>
           <Disclaimer className="hide-at-mobile">
             <p>&copy; 2022 HealthMarkets Insurance Agency. All rights reserved.</p>
-            <p>Attention: This website is operated by HealthMarkets Insurance Agency and is not the Health Insurance Marketplace® website. In offering this website, HealthMarkets Insurance Agency is required to comply with all applicable federal laws, including the standards established under 45 C.F.R. § 155.220(c) and (d) and standards established under 45 C.F.R. § 155.260 to protect the privacy and security of personally identifiable information. This website may not display all data on Qualified Health Plans (QHPs) being offered in your state through the Health Insurance Marketplace® website. To see all available data on Qualified Health Plan options in your state, go to the Health Insurance Marketplace® website at HealthCare.gov.</p>
-            <p>HealthMarkets Insurance Agency offers the opportunity to enroll in either QHPs or off-Marketplace coverage. Please visit HealthCare.gov for information on the benefits of enrolling in a QHP. Off-Marketplace coverage is not eligible for the cost savings offered for coverage through the Marketplaces.</p>
-            <p>HealthMarkets Insurance Agency, Inc. is licensed as an insurance agency in all 50 states and DC. Not all agents are licensed to sell all products. Service and product availability varies by state. Sales agents may be compensated based on a consumer’s enrollment in a health plan. No obligation to enroll. Agent cannot provide tax or legal advice. Contact your tax or legal professional to discuss details regarding your individual business circumstances. Our quoting tool is provided for your information only. All quotes are estimates and are not final until consumer is enrolled. Medicare has neither reviewed nor endorsed this information.</p>
-            <p><strong>Sources:</strong></p>
-            <p>Kaiser Family Foundation (2019). Average Single Premium per Enrolled Employee For Employer-Based Health Insurance.</p>
-            <p>46513-HM-1020</p>
+            {children}
           </Disclaimer>
         </WidgetWrapper>
         <PlansByStateMenu/>
         <Disclaimer className="show-at-mobile">
           <p>&copy; 2022 HealthMarkets Insurance Agency. All rights reserved.</p>
-          <p>Attention: This website is operated by HealthMarkets Insurance Agency and is not the Health Insurance Marketplace® website. In offering this website, HealthMarkets Insurance Agency is required to comply with all applicable federal laws, including the standards established under 45 C.F.R. § 155.220(c) and (d) and standards established under 45 C.F.R. § 155.260 to protect the privacy and security of personally identifiable information. This website may not display all data on Qualified Health Plans (QHPs) being offered in your state through the Health Insurance Marketplace® website. To see all available data on Qualified Health Plan options in your state, go to the Health Insurance Marketplace® website at HealthCare.gov.</p>
-          <p>HealthMarkets Insurance Agency offers the opportunity to enroll in either QHPs or off-Marketplace coverage. Please visit HealthCare.gov for information on the benefits of enrolling in a QHP. Off-Marketplace coverage is not eligible for the cost savings offered for coverage through the Marketplaces.</p>
-          <p>HealthMarkets Insurance Agency, Inc. is licensed as an insurance agency in all 50 states and DC. Not all agents are licensed to sell all products. Service and product availability varies by state. Sales agents may be compensated based on a consumer’s enrollment in a health plan. No obligation to enroll. Agent cannot provide tax or legal advice. Contact your tax or legal professional to discuss details regarding your individual business circumstances. Our quoting tool is provided for your information only. All quotes are estimates and are not final until consumer is enrolled. Medicare has neither reviewed nor endorsed this information.</p>
-          <p><strong>Sources:</strong></p>
-          <p>Kaiser Family Foundation (2019). Average Single Premium per Enrolled Employee For Employer-Based Health Insurance.</p>
-          <p>46513-HM-1020</p>
+          {children}
         </Disclaimer>
       </WidgetAreaTwo>
     </FooterWrapper>
