+++ src/hooks/insurance/useDentalPageQuery.tsx
@@ -54,6 +54,9 @@ export const useDentalPageQuery = () => {
             }
           }
         }
+        disclaimers {
+          disclaimer
+        }
       }
     }
   `)
