+++ src/hooks/insurance/useMedicarePageQuery.tsx
@@ -58,6 +58,9 @@ export const useMedicarePageQuery = () => {
             }
           }
         }
+        disclaimers {
+          disclaimer
+        }
       }
     }
   `)
