+++ src/hooks/insurance/useSmallBusinessPageQuery.tsx
@@ -93,6 +93,9 @@ export const useSmallBusinessPageQuery = () => {
             }
           }
         }
+        disclaimers {
+          disclaimer
+        }
       }
     }
   `)
