+++ src/hooks/useAboutUsPageQuery.tsx
@@ -28,6 +28,9 @@ export const useAboutUsPageQuery = () => {
           }
           gpAboutContent
         }
+        disclaimers {
+          disclaimer
+        }
       }
     }
   `)
