+++ src/hooks/useContactPageQuery.tsx
@@ -1,0 +1,16 @@
+import { useStaticQuery, graphql } from "gatsby";
+
+export const useContactPageQuery = () => {
+  const data = useStaticQuery(graphql`
+    query contactQuery {
+      page: wpPage(databaseId: {eq: 1067}) {
+        id
+        disclaimers {
+          disclaimer
+        }
+      }
+    }
+  `)
+
+  return data
+}
