+++ src/pages/health-insurance/index.tsx
@@ -32,6 +32,7 @@ import Accordion from "../../components/Accordions";
 import Carousel from "../../components/Carousel";
 import Review from "../../components/Reviews/Review";
 import Reviews from "../../components/Reviews";
+import Footer from "../../components/Footer";
 
 const HealthInsurancePage = () => {
   const { page } = useHealthPageQuery();
@@ -174,6 +175,9 @@ const HealthInsurancePage = () => {
           <Button background="accent" border="accent" color="light">View more articles</Button>
         </div>
       </Section>
+      <Footer>
+        <div dangerouslySetInnerHTML={{ __html: page.disclaimers.disclaimer }} />
+      </Footer>
     </Layout>
   )
 }
