+++ src/pages/plans-by-state/dental-insurance/{wpPlansByState.slug}.tsx
@@ -82,6 +82,9 @@ query($slug: String!)
                     }
                 }
             }
+            disclaimers {
+                disclaimer
+            }
         }
     }
 `
