+++ src/pages/plans-by-state/health-insurance/{wpPlansByState.slug}.tsx
@@ -80,6 +80,9 @@ query($slug: String!)
                     }
                 }
             }
+            disclaimers {
+                disclaimer
+            }
         }
     }
 `
