+++ src/pages/plans-by-state/life-insurance/{wpPlansByState.slug}.tsx
@@ -81,6 +81,9 @@ query($slug: String!)
                     }
                 }
             }
+            disclaimers {
+                disclaimer
+            }
         }
     }
 `
